backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
motor==3.4.0
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
